nyuvm/requirements.txt
pillow
django-compressor
django-leaflet
